mesh-1466: 11. extract RT commons
